fix: 불필요한 참조 및 변수 제거
