+++ src/components/markdown/MarkdownLoader.tsx
@@ -9,6 +9,7 @@ import { Header } from "../common/Header";
 import { SlateBlock } from "../common/SlateBlock";
 import { MarkdownLeaf } from "../markdown/MarkdownLeaf";
 import { Element } from "../common/Element";
+
 export const MarkdownLoader = () => {
   const [editor] = useState(() => withReact(createEditor()));
   const fileExtension = ".md";
@@ -28,7 +29,7 @@ export const MarkdownLoader = () => {
     (props: RenderLeafProps) => <MarkdownLeaf {...props} />,
     []
   );
-  const handleChange = useCallback(
+  const parseMarkdownText = useCallback(
     (nextValue: Descendant[]) => {
       setValue(nextValue);
       setMarkdownText(value.map((v) => serialize(v)).join("\n"));
@@ -39,7 +40,7 @@ export const MarkdownLoader = () => {
   return (
     <div css={containerStyle}>
       <Header />
-      <Slate editor={editor} value={initialValue} onChange={handleChange}>
+      <Slate editor={editor} value={initialValue} onChange={parseMarkdownText}>
         <SlateBlock
           fileExtension={fileExtension}
           setStateFunction={setStateFunction}
