+++ src/types/type.ts
@@ -1,3 +1,4 @@
+import JSZip from "jszip";
 import { BaseEditor } from "slate";
 import { ReactEditor } from "slate-react";
 
@@ -13,3 +14,17 @@ declare module "slate" {
 }
 
 export type setStateFunctionType = (value: string) => void;
+
+/* epub parser */
+export interface ICollectedFile {
+  fileName: string;
+  fileExt: string;
+  filePaths: string[];
+  originName: string;
+  file: JSZip.JSZipObject;
+}
+
+export interface ICollectedData {
+  containerXML: string;
+  files: ICollectedFile[];
+}
